Include CSS rules to style the basic HTML page
I used this exercise to practice using class and
ID selectors on the provided div and p HTML
elements.

The solution mentioned a spot in the CSS file
where the developer can reduce the amount of
duplicate code by mixing the use of class and ID
HTML attributes, which was a nice reminder!