Arrange the page images and text to desired format
I used this exercise to arrange the plant images
and text to the requirements described in its
respective README file. This lesson's intent to
apply the developer's acquired knowledge of
flexbox properties in the module.

I learned that applying the text-align property
will align the image and the text inside of a div
element if they are both inside of a container
div element. That is a more simple solution than
what I programmed in this exercise, and will be
noted in the future.